added cart logic, 401 error need to fix

+++ PizzaToPizza/Dtos/PromoCodes/UpdatePromoCodeDto.cs
@@ -3,7 +3,7 @@
     public class UpdatePromoCodeDto
     {
         public string Code { get; set; } = "";
-        public decimal DiscountPercent { get; set; }
+        public int DiscountPercent { get; set; }
         public DateTime ExpiryDate { get; set; }
     }
 }
